feat: expose a billing service at api
This endpoint is a private/admin endpoint and can not be exposed api gateway.
I decided that to be easier to trigger the billing process manually if it necessary.
This endpoint also will be used by the AWS Event Bridge that will schedule a request execution on first day week of the month.